Refactory WebApi.Tests to UdapMetadata.Tests
