+++ .github/workflows/dotnet.yml
@@ -31,8 +31,8 @@ jobs:
       run: dotnet build -c Release Udap.sln
     - name: Test Udap.Common.Tests
       run: dotnet test -c Release _tests/Udap.Common.Tests/Udap.Common.Tests.csproj --no-build --verbosity normal
-    - name: Test WebApi.Tests
-      run: dotnet test -c Release _tests/WebApi.Tests/WebApi.Tests.csproj --no-build --verbosity normal
+    - name: Test UdapMetadata.Tests
+      run: dotnet test -c Release _tests/UdapMetadata.Tests/UdapMetadata.Tests.csproj --no-build --verbosity normal
     - name: Test UdapServer.Tests
       run: dotnet test -c Release _tests/UdapServer.Tests/UdapServer.Tests.csproj --no-build --verbosity normal
