+++ Udap.sln
@@ -31,8 +31,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Udap.Idp", "examples\Udap.I
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Udap.Idp.Admin", "examples\Udap.Idp.Admin\Udap.Idp.Admin.csproj", "{AE71744D-445C-443F-A903-36F895CFDDD5}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebApi.Tests", "_tests\WebApi.Tests\WebApi.Tests.csproj", "{8715D597-944E-417E-8DDC-AB57DCA6477C}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Udap.Common.Tests", "_tests\Udap.Common.Tests\Udap.Common.Tests.csproj", "{3AA40005-F411-43E6-8503-37AE8350F84F}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Udap.Support.Tests", "_tests\Udap.Support.Tests\Udap.Support.Tests.csproj", "{1B6A6580-E21F-4E4B-AC78-56728385D3B8}"
@@ -62,6 +60,8 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Udap.CA.Tests", "_tests\Udap.CA.Tests\Udap.CA.Tests.csproj", "{9191CC9B-4DD0-4115-8FFD-A710659DBC27}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Udap.Client.System.Tests", "_tests\Udap.Client.System.Tests\Udap.Client.System.Tests.csproj", "{63DEA604-5D81-4C3A-AF43-0A296372EB20}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "UdapMetadata.Tests", "_tests\UdapMetadata.Tests\UdapMetadata.Tests.csproj", "{872BDB24-9C83-4089-A5ED-086CAC6103B1}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -101,10 +101,6 @@ Global
 		{AE71744D-445C-443F-A903-36F895CFDDD5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{AE71744D-445C-443F-A903-36F895CFDDD5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{AE71744D-445C-443F-A903-36F895CFDDD5}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8715D597-944E-417E-8DDC-AB57DCA6477C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8715D597-944E-417E-8DDC-AB57DCA6477C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8715D597-944E-417E-8DDC-AB57DCA6477C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8715D597-944E-417E-8DDC-AB57DCA6477C}.Release|Any CPU.Build.0 = Release|Any CPU
 		{3AA40005-F411-43E6-8503-37AE8350F84F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{3AA40005-F411-43E6-8503-37AE8350F84F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3AA40005-F411-43E6-8503-37AE8350F84F}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -157,6 +153,10 @@ Global
 		{63DEA604-5D81-4C3A-AF43-0A296372EB20}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{63DEA604-5D81-4C3A-AF43-0A296372EB20}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{63DEA604-5D81-4C3A-AF43-0A296372EB20}.Release|Any CPU.Build.0 = Release|Any CPU
+		{872BDB24-9C83-4089-A5ED-086CAC6103B1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{872BDB24-9C83-4089-A5ED-086CAC6103B1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{872BDB24-9C83-4089-A5ED-086CAC6103B1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{872BDB24-9C83-4089-A5ED-086CAC6103B1}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -166,7 +166,6 @@ Global
 		{B05D207B-3354-49A8-ABDE-037E81A1F4F0} = {4B854644-1F54-4ECA-9290-A89B91A19FF9}
 		{E4027C2E-4502-4D6E-97EB-1E5F7C251148} = {4B854644-1F54-4ECA-9290-A89B91A19FF9}
 		{AE71744D-445C-443F-A903-36F895CFDDD5} = {4B854644-1F54-4ECA-9290-A89B91A19FF9}
-		{8715D597-944E-417E-8DDC-AB57DCA6477C} = {640C178B-FA29-49AD-8245-FA7EE934D522}
 		{3AA40005-F411-43E6-8503-37AE8350F84F} = {640C178B-FA29-49AD-8245-FA7EE934D522}
 		{1B6A6580-E21F-4E4B-AC78-56728385D3B8} = {640C178B-FA29-49AD-8245-FA7EE934D522}
 		{DA8D6B73-3C01-4FDE-8822-97943443AC24} = {640C178B-FA29-49AD-8245-FA7EE934D522}
@@ -180,6 +179,7 @@ Global
 		{A634BD23-B31B-4260-922E-24FEC8B4CF22} = {4B854644-1F54-4ECA-9290-A89B91A19FF9}
 		{9191CC9B-4DD0-4115-8FFD-A710659DBC27} = {640C178B-FA29-49AD-8245-FA7EE934D522}
 		{63DEA604-5D81-4C3A-AF43-0A296372EB20} = {640C178B-FA29-49AD-8245-FA7EE934D522}
+		{872BDB24-9C83-4089-A5ED-086CAC6103B1} = {640C178B-FA29-49AD-8245-FA7EE934D522}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {4C22E2C0-B5D3-497F-BAD0-EB769D6D6CD1}
