+++ _tests/UdapMetadata.Tests/WeatherApi/UdapControllerTests.cs
@@ -20,30 +20,27 @@ using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Microsoft.IdentityModel.Tokens;
 using Newtonsoft.Json;
-using Udap.Common;
-using Udap.Metadata.Server;
 using Udap.Model;
 using Xunit.Abstractions;
-using program = FhirLabsApi.Program;
+using program = WeatherApi.Program;
 
+namespace UdapMetadata.Tests.WeatherApi;
 
-namespace WebApi.Tests.FhirLabsApi;
-
-public class ApiTestFixture : WebApplicationFactory<program>
+public class ApiTestFixture : WebApplicationFactory<program> 
 {
-    private UdapMetadata? _wellKnownUdap;
-    public ITestOutputHelper? Output { get; set; }
-    
-    public UdapMetadata WellKnownUdap
+    private Udap.Model.UdapMetadata? _wellKnownUdap;
+    public ITestOutputHelper Output { get; set; } = null!;
+
+    public Udap.Model.UdapMetadata WellKnownUdap
     {
         get
         {
             if (_wellKnownUdap == null)
             {
-                var response = CreateClient().GetAsync($"fhir/r4/.well-known/udap").GetAwaiter().GetResult();
+                var response = CreateClient().GetAsync(".well-known/udap").GetAwaiter().GetResult();
                 response.StatusCode.Should().Be(HttpStatusCode.OK);
                 var content = response.Content.ReadAsStringAsync().GetAwaiter().GetResult();
-                _wellKnownUdap = System.Text.Json.JsonSerializer.Deserialize<UdapMetadata>(content);
+                _wellKnownUdap = System.Text.Json.JsonSerializer.Deserialize<Udap.Model.UdapMetadata>(content);
             }
 
             return _wellKnownUdap!;
@@ -57,7 +54,7 @@ public class ApiTestFixture : WebApplicationFactory<program>
         // Still works with Windows but what a pain.  This feels fragile
         // TODO: 
         //
-        builder.UseSetting("contentRoot", "../../../../../examples/FhirLabsApi");
+        builder.UseSetting("contentRoot", "../../../../../examples/WeatherApi");
     }
 
     protected override IHost CreateHost(IHostBuilder builder)
@@ -66,7 +63,7 @@ public class ApiTestFixture : WebApplicationFactory<program>
         builder.ConfigureLogging(logging =>
         {
             logging.ClearProviders();
-            logging.AddXUnit(Output!);
+            logging.AddXUnit(Output);
         });
 
         return base.CreateHost(builder);
@@ -75,9 +72,10 @@ public class ApiTestFixture : WebApplicationFactory<program>
 
 public class UdapControllerTests : IClassFixture<ApiTestFixture>
 {
-    private readonly ApiTestFixture _fixture;
+    private ApiTestFixture _fixture;
+    private readonly ITestOutputHelper _testOutputHelper;
 
-    public UdapControllerTests(ApiTestFixture fixture, ITestOutputHelper output)
+    public UdapControllerTests(ApiTestFixture fixture, ITestOutputHelper output, ITestOutputHelper testOutputHelper)
     {
         //
         // Tests json once
@@ -85,6 +83,7 @@ public class UdapControllerTests : IClassFixture<ApiTestFixture>
         if (fixture == null) throw new ArgumentNullException(nameof(fixture));
         fixture.Output = output;
         _fixture = fixture;
+        _testOutputHelper = testOutputHelper;
     }
 
     /// <summary>
@@ -103,9 +102,8 @@ public class UdapControllerTests : IClassFixture<ApiTestFixture>
     [Fact]
     public void udap_versions_supportedTest()
     {
-        var verSupported = _fixture.WellKnownUdap.UdapVersionsSupported;
-        verSupported.Should().NotBeNullOrEmpty();
-        verSupported!.Single().Should().Be("1");
+        var udapVerSupported = _fixture.WellKnownUdap.UdapVersionsSupported?.SingleOrDefault();
+        udapVerSupported.Should().Be("1");
     }
 
 
@@ -117,6 +115,9 @@ public class UdapControllerTests : IClassFixture<ApiTestFixture>
 
         var hl7B2B = extensions!.SingleOrDefault(c => c == "hl7-b2b");
         hl7B2B.Should().NotBeNullOrEmpty();
+
+        var acmeExt = extensions!.SingleOrDefault(c => c == "acme-ext");
+        acmeExt.Should().NotBeNullOrEmpty();
     }
 
     /// <summary>
@@ -151,8 +152,7 @@ public class UdapControllerTests : IClassFixture<ApiTestFixture>
     [Fact]
     public void udap_certifications_requiredTest()
     {
-        var certificationsSupported = _fixture.WellKnownUdap.UdapCertificationsRequired?.SingleOrDefault();
-        certificationsSupported.Should().NotBeNullOrEmpty();
+        var certificationsSupported = _fixture.WellKnownUdap.UdapCertificationsRequired?.Single();
         var uriCertificationsSupported = new Uri(certificationsSupported!);
         uriCertificationsSupported.Should().Be("http://MyUdapCertification");
     }
@@ -173,7 +173,6 @@ public class UdapControllerTests : IClassFixture<ApiTestFixture>
     public void scopes_supported_supportedTest()
     {
         var scopesSupported = _fixture.WellKnownUdap.ScopesSupported;
-
         scopesSupported.Should().Contain("openid");
         scopesSupported.Should().Contain("system/Patient.read");
         scopesSupported.Should().Contain("system/AllergyIntolerance.read");
@@ -184,44 +183,41 @@ public class UdapControllerTests : IClassFixture<ApiTestFixture>
     public void authorization_endpointTest()
     {
         var authorizationEndpoint = _fixture.WellKnownUdap.AuthorizationEndpoint;
-        authorizationEndpoint.Should().Be("https://host.docker.internal:5002/connect/authorize");
+        authorizationEndpoint.Should().Be("https://securedcontrols.net:5001/connect/authorize");
     }
 
     [Fact]
     public void token_endpointTest()
     {
         var tokenEndpoint = _fixture.WellKnownUdap.TokenEndpoint;
-        tokenEndpoint.Should().Be("https://host.docker.internal:5002/connect/token");
+        tokenEndpoint.Should().Be("https://securedcontrols.net:5001/connect/token");
     }
 
     [Fact]
     public void registration_endpointTest()
     {
         var registrationEndpoint = _fixture.WellKnownUdap.RegistrationEndpoint;
-        registrationEndpoint.Should().Be("https://host.docker.internal:5002/connect/register");
+        registrationEndpoint.Should().Be("https://securedcontrols.net:5001/connect/register");
     }
 
     [Fact]
     public void token_endpoint_auth_methods_supportedTest()
     {
-        var scopesSupported = _fixture.WellKnownUdap.TokenEndpointAuthMethodsSupported?.SingleOrDefault();
-        scopesSupported.Should().NotBeNullOrEmpty();
+        var scopesSupported = _fixture.WellKnownUdap.TokenEndpointAuthMethodsSupported?.Single();
         scopesSupported.Should().Be("private_key_jwt");
     }
 
     [Fact]
     public void token_endpoint_auth_signing_alg_values_supportedTest()
     {
-        var scopesSupported = _fixture.WellKnownUdap.TokenEndpointAuthSigningAlgValuesSupported?.SingleOrDefault();
-        scopesSupported.Should().NotBeNullOrEmpty();
+        var scopesSupported = _fixture.WellKnownUdap.TokenEndpointAuthSigningAlgValuesSupported?.Single();
         scopesSupported.Should().Be(UdapConstants.SupportedAlgorithm.RS256);
     }
 
     [Fact]
     public void registration_endpoint_jwt_signing_alg_values_supportedTest()
     {
-        var scopesSupported = _fixture.WellKnownUdap.RegistrationEndpointJwtSigningAlgValuesSupported?.SingleOrDefault();
-        scopesSupported.Should().NotBeNullOrEmpty();
+        var scopesSupported = _fixture.WellKnownUdap.RegistrationEndpointJwtSigningAlgValuesSupported?.Single();
         scopesSupported.Should().Be(UdapConstants.SupportedAlgorithm.RS256);
     }
 
@@ -229,7 +225,6 @@ public class UdapControllerTests : IClassFixture<ApiTestFixture>
     public void signed_metadataTest()
     {
         var signedMetatData = _fixture.WellKnownUdap.SignedMetadata;
-        signedMetatData.Should().NotBeNullOrEmpty();
 
         var pattern = @"^[A-Za-z0-9-_=]+\.[A-Za-z0-9-_=]+\.?[A-Za-z0-9-_.+\/=]*$";
         var regex = new Regex(pattern);
@@ -237,20 +232,18 @@ public class UdapControllerTests : IClassFixture<ApiTestFixture>
     }
 
     [Fact]
-    public void signed_metadataContentTest()
+    public void signed_metatdataContentTest()
     {
         var jwt = new JwtSecurityToken(_fixture.WellKnownUdap.SignedMetadata);
         var tokenHeader = jwt.Header;
-        
         var x5CArray = JsonConvert.DeserializeObject<string[]>(tokenHeader.X5c);
+        x5CArray.Should().NotBeNull();
 
         // bad keys
         //x5cArray[0] = "MIIFJDCCBAygAwIBAgIIUFnObaPiufEwDQYJKoZIhvcNAQELBQAwgbMxCzAJBgNVBAYTAlVTMRMwEQYDVQQIDApDYWxpZm9ybmlhMRIwEAYDVQQHDAlTYW4gRGllZ28xEzARBgNVBAoMCkVNUiBEaXJlY3QxPzA9BgNVBAsMNlRlc3QgUEtJIENlcnRpZmljYXRpb24gQXV0aG9yaXR5IChjZXJ0cy5lbXJkaXJlY3QuY29tKTElMCMGA1UEAwwcRU1SIERpcmVjdCBUZXN0IENsaWVudCBTdWJDQTAeFw0yMTAxMTUyMTQ1MTRaFw0yNDAxMTYyMTQ1MTRaMIGlMQswCQYDVQQGEwJVUzETMBEGA1UECAwKQ2FsaWZvcm5pYTETMBEGA1UECgwKRU1SIERpcmVjdDEzMDEGA1UECwwqVURBUCBUZXN0IENlcnRpZmljYXRlIE5PVCBGT1IgVVNFIFdJVEggUEhJMTcwNQYDVQQDDC5odHRwczovL3N0YWdlLmhlYWx0aHRvZ28ubWU6ODE4MS9maGlyL3I0L3N0YWdlMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAt9j718Yu8HjoIdSvLTloVLnFLdfdL7T/BylPcIpcKhB7zJvNzZOpq8T/fXhc9b4p6cY6gBPBq1Vnax4zTCAP/te5W6FfoRoKhKqpExuYmgIw0lE8a4UAnHVwPOAvuKS3abGzYfLxxUc4PFXp4HrBx/QWOMqR408GlbSYG0wpeifhMx1VD8TFmU13FmFqgP3cEHjT7RxulfJnPcPPXZ8b5tZIkQMlApJRULVnHEBcICixaRWCJjzzArgoFUydPiAfMZELi80W4n0Wn/WduSYZqwQAosI7AfS3NINd44w8kek1X9WVwX/QtcAVuCXvSFoqoIAa3l4kBCQIHmY9UhltZwIDAQABo4IBRjCCAUIwWQYIKwYBBQUHAQEETTBLMEkGCCsGAQUFBzAChj1odHRwOi8vY2VydHMuZW1yZGlyZWN0LmNvbS9jZXJ0cy9FTVJEaXJlY3RUZXN0Q2xpZW50U3ViQ0EuY3J0MB0GA1UdDgQWBBRZmXqpQzFDSamfvPKiKtjg9gp8cTAMBgNVHRMBAf8EAjAAMB8GA1UdIwQYMBaAFKOVbWu9K1HN4c/lkG/XJk+/3T7eMEwGA1UdHwRFMEMwQaA/oD2GO2h0dHA6Ly9jZXJ0cy5lbXJkaXJlY3QuY29tL2NybC9FTVJEaXJlY3RUZXN0Q2xpZW50U3ViQ0EuY3JsMA4GA1UdDwEB/wQEAwIHgDA5BgNVHREEMjAwhi5odHRwczovL3N0YWdlLmhlYWx0aHRvZ28ubWU6ODE4MS9maGlyL3I0L3N0YWdlMA0GCSqGSIb3DQEBCwUAA4IBAQAePi+wIAPubt2Fk2jbELZt/bgkc7KTGC5C4sLX25NNYyzvHh0kwmHvgBx3thCv7uOvf/nbmhnk+l3EmgdaB1ZjzcjLMFc7xec9YJWsegzEkR2pzYQp/41cmhTfwNSnXxUSZrBtqInx+mALi9r96lg6RpqQh+DxlToC2vreW7Fy3pFa3DQKFN6j6azYTj5ljqrGprKQRh/iyqRvY+j+BC44Wl+POfBVObwtf71irMuLsSCmMptPGFGTqQdtLYbFjkB4wowiFfEe0PYL+N015iPZA4wimlXbau4XaEvipnIsWxqzT30RbQgrrOw7zN1QjGRURBbdBkMrgLkzmfGxhjuV";
 
         var cert = new X509Certificate2(Convert.FromBase64String(x5CArray!.First()));
-
         var tokenHandler = new JwtSecurityTokenHandler();
-
 
         tokenHandler.ValidateToken(_fixture.WellKnownUdap.SignedMetadata, new TokenValidationParameters
         {
@@ -264,12 +257,12 @@ public class UdapControllerTests : IClassFixture<ApiTestFixture>
         var issClaim = jwt.Payload.Claims.Single(c => c.Type == JwtClaimTypes.Issuer);
         issClaim.ValueType.Should().Be(ClaimValueTypes.String);
 
-        // should be the same as the web base url
-        issClaim.Value.Should().Be("https://fhirlabs.net:7016/fhir/r4");
+        // should be the same as the web base url, but this would be localhost
+        issClaim.Value.Should().Be("https://weatherapi.lab:5021/fhir");
 
         var subjectAltName = cert.GetNameInfo(X509NameType.UrlName, false);
         subjectAltName.Should().Be(issClaim.Value, $"iss: {issClaim.Value} does not match Subject Alternative Name extension");
-        
+
         var subClaim = jwt.Payload.Claims.Single(c => c.Type == JwtClaimTypes.Subject);
         subClaim.ValueType.Should().Be(ClaimValueTypes.String);
 
@@ -286,5 +279,7 @@ public class UdapControllerTests : IClassFixture<ApiTestFixture>
         var exp = int.Parse(expClaim.Value);
         var year = DateTimeOffset.FromUnixTimeSeconds(exp).AddYears(1).ToUnixTimeSeconds();
         iat.Should().BeLessOrEqualTo((int)year);
+
+
     }
 }
